Extract DispatchBinding into ts-helpers package
