+++ core/chainlink-sgx.Dockerfile
@@ -17,7 +17,7 @@ COPY styleguide/package.json ./styleguide/
 COPY tools/json-api-client/package.json ./tools/json-api-client/
 COPY tools/local-storage/package.json ./tools/local-storage/
 COPY tools/redux/package.json ./tools/redux/
-COPY tools/ts-test-helpers/package.json ./tools/ts-test-helpers/
+COPY tools/ts-helpers/package.json ./tools/ts-helpers/
 COPY belt/package.json ./belt/
 COPY belt/bin ./belt/bin
 COPY evm-test-helpers/package.json ./evm-test-helpers/
@@ -37,7 +37,7 @@ COPY styleguide ./styleguide
 COPY tools/json-api-client ./tools/json-api-client
 COPY tools/local-storage ./tools/local-storage
 COPY tools/redux ./tools/redux
-COPY tools/ts-test-helpers ./tools/ts-test-helpers
+COPY tools/ts-helpers ./tools/ts-helpers
 COPY belt ./belt
 COPY belt/bin ./belt/bin
 COPY evm-test-helpers ./evm-test-helpers
