+++ explorer/client/src/__tests__/reducers/jobRuns.test.ts
@@ -1,4 +1,4 @@
-import { partialAsFull } from '@chainlink/ts-test-helpers'
+import { partialAsFull } from '@chainlink/ts-helpers'
 import { JobRun } from 'explorer/models'
 import reducer, {
   INITIAL_STATE as initialRootState,
