+++ explorer/client/src/__tests__/reducers/jobRunsIndex.test.ts
@@ -1,4 +1,4 @@
-import { partialAsFull } from '@chainlink/ts-test-helpers'
+import { partialAsFull } from '@chainlink/ts-helpers'
 import reducer, {
   INITIAL_STATE as initialRootState,
   AppState,
