+++ explorer/client/src/__tests__/utils/status.test.ts
@@ -1,4 +1,4 @@
-import { partialAsFull } from '@chainlink/ts-test-helpers'
+import { partialAsFull } from '@chainlink/ts-helpers'
 import status from '../../utils/status'
 import { JobRun, TaskRun } from 'explorer/models'
