+++ explorer/client/src/containers/Admin/Operator/Index.tsx
@@ -10,12 +10,12 @@ import { ChainlinkNode } from 'explorer/models'
 import React, { useEffect, useState } from 'react'
 import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux'
 import build from 'redux-object'
+import { DispatchBinding } from '@chainlink/ts-helpers'
 import { fetchAdminOperators } from '../../../actions/adminOperators'
 import List from '../../../components/Admin/Operators/List'
 import { ChangePageEvent } from '../../../components/Table'
 import Title from '../../../components/Title'
 import { AppState } from '../../../reducers'
-import { DispatchBinding } from '../../../utils/types'
 
 const styles = ({ breakpoints, spacing }: Theme) =>
   createStyles({
