+++ explorer/client/src/containers/Admin/SignIn.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux'
 import { Redirect, RouteComponentProps } from '@reach/router'
+import { DispatchBinding } from '@chainlink/ts-helpers'
 import { SignIn as SignInForm } from '../../components/Forms/SignIn'
 import { signIn } from '../../actions/adminAuth'
 import { AppState } from '../../reducers'
-import { DispatchBinding } from '../../utils/types'
 
 interface OwnProps {}
