+++ explorer/client/src/containers/JobRuns/Index.tsx
@@ -4,15 +4,15 @@ import {
   withStyles,
   WithStyles,
 } from '@material-ui/core/styles'
-import { JobRun } from 'explorer/models'
 import React, { useEffect, useState } from 'react'
 import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux'
 import build from 'redux-object'
+import { DispatchBinding } from '@chainlink/ts-helpers'
+import { JobRun } from 'explorer/models'
 import { fetchJobRuns } from '../../actions/jobRuns'
 import List from '../../components/JobRuns/List'
 import { ChangePageEvent } from '../../components/Table'
 import { AppState } from '../../reducers'
-import { DispatchBinding } from '../../utils/types'
 
 const EMPTY_MSG =
   "We couldn't find any results for your search query. Try again with the job id, run id, requester, requester id or transaction hash"
