+++ explorer/client/tsconfig.json
@@ -10,15 +10,9 @@
   },
   "include": ["src", "@types", "test"],
   "references": [
-    {
-      "path": "../../tools/json-api-client"
-    },
-    {
-      "path": "../../tools/redux"
-    },
-    {
-      "path": "../../tools/ts-test-helpers"
-    },
+    { "path": "../../tools/json-api-client" },
+    { "path": "../../tools/redux" },
+    { "path": "../../tools/ts-helpers" },
     { "path": "../../styleguide" }
   ]
 }
