+++ operator_ui/__tests__/actions.test.ts
@@ -1,4 +1,4 @@
-import { partialAsFull } from '@chainlink/ts-test-helpers'
+import { partialAsFull } from '@chainlink/ts-helpers'
 import * as models from 'core/store/models'
 import { bindActionCreators, Middleware } from 'redux'
 import configureStore from 'redux-mock-store'
