+++ operator_ui/__tests__/selectors/job.test.ts
@@ -1,4 +1,4 @@
-import { partialAsFull } from '@chainlink/ts-test-helpers'
+import { partialAsFull } from '@chainlink/ts-helpers'
 import { INITIAL_STATE, AppState } from '../../src/reducers'
 import jobSelector from '../../src/selectors/job'
