+++ package.json
@@ -19,7 +19,7 @@
     "tools/json-api-client",
     "tools/local-storage",
     "tools/redux",
-    "tools/ts-test-helpers",
+    "tools/ts-helpers",
     "belt"
   ],
   "engines": {
