+++ tools/ts-helpers/src/index.ts
@@ -1,1 +1,2 @@
 export * from './mocks'
+export * from './types'
